fix: consolidate tile API route
